AM-2857: Clean up controllers
